Move vocab-action mixin to a separate directory.
