+++ app/mixins/vocab-action.js
